# docs/api-documentation.yml
openapi: 3.0.3
info:
  title: Rent360 API
  description: API completa para la plataforma de gestión inmobiliaria Rent360
  version: 1.0.0
  contact:
    name: Rent360 Support
    email: support@rent360.cl
  license:
    name: MIT
servers:
  - url: https://api.rent360.cl/v1
    description: Production server
  - url: http://localhost:3000/api
    description: Development server
security:
  - bearerAuth: []
paths:
...
  /auth/login:
    post:
      summary: Iniciar sesión
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
      responses:
        200:
          description: Login exitoso
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
        401:
          description: Credenciales inválidas
...
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        name:
          type: string
        role:
          type: string
          enum: [ADMIN, OWNER, TENANT, BROKER, RUNNER, SUPPORT]
        isActive:
          type: boolean
        avatar:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT